Improve column views loading
