+++ packages/form-components/package.json
@@ -35,6 +35,7 @@
     "@types/mapbox__mapbox-gl-draw": "^1.2.3",
     "@types/react-dom": "^17.0.16||^18.0.0",
     "axios": "^1.7.9",
+    "classnames": "^2.5.1",
     "mapbox-gl": "^2.15.0"
   },
   "type": "module",
